github_api/requirements.txt
requests==2.31.0
python-dotenv==1.2.1
pymongo==4.16.0